Move register -> GX_Material parsing to gx_material.ts
No functional changes to brres
